Big Make a List changes
Big changes to list making. Added functionality to allow searches to
query lists of variables (like multiple zip codes) and also so that
searches can accept as many or as few variables as wanted.

Also added a nav menu for logged in users and a way to log out

TODO:
-This new way is not as secure. NEED to check and make sure we are not
opening up to SQL Inject.
-export functionality
-Time users out